Refactor: Use a more robust stream management
This commit refactors the stream management to be more robust. It
includes: - Adding a ready channel to synchronize download and upload -
Returning 404 if the download handler is called before upload -
Implementing a timeout for the upload handler to wait for the download
client. - Adding better error handling to the upload/download streams -
Correcting the body type of the download handler